Move Settings model to storage/data_models and update imports
