+++ openhands/storage/settings/file_settings_store.py
@@ -4,7 +4,7 @@ import json
 from dataclasses import dataclass
 
 from openhands.core.config.app_config import AppConfig
-from openhands.server.settings import Settings
+from openhands.storage.data_models.settings import Settings
 from openhands.storage import get_file_store
 from openhands.storage.files import FileStore
 from openhands.storage.settings.settings_store import SettingsStore
